Added comment in lint rules file and refactored folder structure
